+++ exonum/src/encoding/serialize/json.rs
@@ -16,9 +16,10 @@
 /// json format.
 ///
 
-// TODO refer to difference between json serialization and exonum_json
-// TODO implement Field for float
-// TODO remove WriteBufferWraper hack (after refactor storage), should be moved into storage
+// TODO refer to difference between json serialization and exonum_json (ECR-156).
+// TODO implement Field for float (ECR-153).
+// TODO remove WriteBufferWraper hack (after refactor storage),
+// should be moved into storage (ECR-156).
 
 use serde_json::value::Value;
 use bit_vec::BitVec;
@@ -32,7 +33,7 @@ use std::error::Error;
 use crypto::{Hash, PublicKey, Signature};
 use helpers::{Height, Round, ValidatorId};
 
-// TODO: should we implement serilize for: `SecretKey`, `Seed` ?
+// TODO: should we implement serialize for: `SecretKey`, `Seed` (ECR-156)?
 
 use encoding::{Field, Offset};
 use super::HexValue;
@@ -344,8 +345,8 @@ where
     }
 }
 
-//\TODO remove `ExonumJsonDeserialize` needs,
-// after it remove impl `ExonumJsonDeserialize` for all types expect struct
+// TODO remove `ExonumJsonDeserialize` needs
+// after it remove impl `ExonumJsonDeserialize` for all types expect struct (ECR-156)
 impl<T> ExonumJson for Vec<T>
 where
     T: ExonumJsonDeserialize + ExonumJson,
@@ -412,7 +413,7 @@ impl ExonumJson for BitVec {
     }
 }
 
-// TODO: Make a macro for tuple struct typedefs?
+// TODO: Make a macro for tuple struct typedefs (ECR-154)?
 impl ExonumJson for Height {
     fn deserialize_field<B: WriteBufferWrapper>(
         value: &Value,