+++ exonum/src/lib.rs
@@ -62,11 +62,14 @@ extern crate tokio_timer;
 extern crate tokio_core;
 extern crate tokio_io;
 extern crate tokio_retry;
+extern crate os_info;
 
 #[macro_use]
 pub mod encoding;
 #[macro_use]
 pub mod messages;
+#[macro_use]
+pub mod helpers;
 pub mod crypto;
 #[doc(hidden)]
 pub mod events;
@@ -74,5 +77,4 @@ pub mod node;
 pub mod storage;
 pub mod blockchain;
 pub mod explorer;
-pub mod helpers;
 pub mod api;