+++ exonum/src/node/state.rs
@@ -32,21 +32,21 @@ use helpers::{Height, Round, ValidatorId, Milliseconds};
 use node::whitelist::Whitelist;
 use node::timeout_adjuster::{TimeoutAdjuster, Constant, Dynamic, MovingAverage};
 
-// TODO: move request timeouts into node configuration
+// TODO: move request timeouts into node configuration (ECR-171)
 
-/// Timeout value for the `RequestPropose` message.
+/// Timeout value for the `ProposeRequest` message.
 pub const PROPOSE_REQUEST_TIMEOUT: Milliseconds = 100;
-/// Timeout value for the `RequestTransactions` message.
+/// Timeout value for the `TransactionsRequest` message.
 pub const TRANSACTIONS_REQUEST_TIMEOUT: Milliseconds = 100;
-/// Timeout value for the `RequestPrevotes` message.
+/// Timeout value for the `PrevotesRequest` message.
 pub const PREVOTES_REQUEST_TIMEOUT: Milliseconds = 100;
-/// Timeout value for the `RequestBlock` message.
+/// Timeout value for the `BlockRequest` message.
 pub const BLOCK_REQUEST_TIMEOUT: Milliseconds = 100;
 
 /// Transactions pool.
-// TODO replace by persistent TxPool
+// TODO replace by persistent TxPool (ECR-171)
 pub type TxPool = Arc<RwLock<BTreeMap<Hash, Box<Transaction>>>>;
-// TODO: reduce copying of Hash
+// TODO: reduce copying of Hash (ECR-171)
 
 /// State of the `NodeHandler`.
 #[derive(Debug)]
@@ -108,13 +108,13 @@ pub struct ValidatorState {
 /// translated to the corresponding request-message.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum RequestData {
-    /// Represents `RequestPropose` message.
+    /// Represents `ProposeRequest` message.
     Propose(Hash),
-    /// Represents `RequestTransactions` message.
+    /// Represents `TransactionsRequest` message.
     Transactions(Hash),
-    /// Represents `RequestPrevotes` message.
+    /// Represents `PrevotesRequest` message.
     Prevotes(Round, Hash),
-    /// Represents `RequestBlock` message.
+    /// Represents `BlockRequest` message.
     Block(Height),
 }
 
@@ -681,7 +681,7 @@ impl State {
                     .remove(&hash);
             }
         }
-        // TODO: destruct/construct structure HeightState instead of call clear
+        // TODO: destruct/construct structure HeightState instead of call clear (ECR-171)
         self.blocks.clear();
         self.proposes.clear();
         self.prevotes.clear();
@@ -689,7 +689,7 @@ impl State {
         if let Some(ref mut validator_state) = self.validator_state {
             validator_state.clear();
         }
-        self.requests.clear(); // FIXME: clear all timeouts
+        self.requests.clear(); // FIXME: clear all timeouts (ECR-171)
     }
 
     /// Returns a list of queued consensus messages.
@@ -976,7 +976,7 @@ impl State {
             match self.validator_state {
                 Some(ref validator_state) => {
                     if let Some(msg) = validator_state.our_prevotes.get(&round) {
-                        // TODO: unefficient
+                        // TODO: unefficient (ECR-171)
                         if Some(*msg.propose_hash()) != self.locked_propose {
                             return true;
                         }